refactor: add cli module
